Från: 	Brad House <brad@mainstreetsoftworks.com> Till: 	buildroot@uclibc.org Ämne: 	[Buildroot] [patch] libxml2 install fails because of missing directory Datum: 	Tue, 14 Aug 2007 16:34:34 -0400  (22.34 CEST)
If libxml2 is installed before $(TARGET_DIR)/usr/lib is created,
it fails because it can't copy the library over.

Also, the dependency tree is a bit messed up.  libxml2.a should
only be installed if target-headers is specified, it's 4MB
on my system and the only .a file I found in my TARGET_DIR
(target-headers was not specified).

The attached patch fixes the above problems.

-Brad