haserl: fix shared cache interaction issue
The ac_cv_path_install override is needed to strip the target binary for
haserl 0.8.0, but it doesn't interact nicely with the shared cache, so
disable the cache for now.

Longer term we should probably forget about install-strip and do the
strip/cleanup/mklibs stuff as a post processing step before the target
filesystem rules are run.