Previous vtun was version 2.6. This does not buld because its "configure" requires that -llzo contains "lzolx_decompress". "vtun" does not build, evenm if liblzo is available.
The LZO package currently used by buildroot
does not contain ANY reference to "lzolx_decompress"

"vtun" has been upgraded to 3.0.2 and now builds OK,
but is yet to be tested on a target.
The previous patch containing three diffs,
has been broken up into three files.

The second patch fails.

This patch tries to replace a perl script
($(VTUN_DIR)/scripts/vtund.rc.debian)
with a shell script with the same name.

In vtun-3.0.2, vtund.rc.debian is a shell script which is
fairly similar to the shell script provided by the patch.
For now, it has been decided not to replace this shell 
script with the script generated by the patch for 2.6

vtun will thus be built with the 3.0.2 vtund.rc.debian.
The start-stop-daemon parameters and other things
in this script may be inappropriate for something based on busybox.

I will leave the decision which script to use,
the 2-6 script or the 3.0.2 script to someone else.

Both files are kept in the directory, but should
be removed once it has been decided what to do
about this script.

Signed-Off by: Ulf Samuelsson <ulf.samuelsson@atmel.com>