php: don't use DOWNLOAD macro
php.net only offers downloads though a strange php script with the file name
embedded in the *MIDDLE* of the URL, which isn't compatible with the DOWNLOAD
macro. Fix it by reverting php.mk hunk of r24689 to go back to $(WGET).

This ofcourse means that the primary site / fallback mirror stuff isn't used.