Add the shared-mime-info package
The shared-mime-info package contains the core database of common
types and the update-mime-database command used to extend it.

The only thing that is actually needed in the target is the file
mime.cache. In order to generate that file, shared-mime-info is
installed into the staging-dir. The idea is that other packages might
add more mime-info files and these should be collected in
staging-dir. From there the mime.cache file is created using tools
installed in host-dir and copied to target.

From Sven Neumann <s.neumann@raumfeld.com>

Modifications by Thomas Petazzoni :

 * Add a patch to the main Makefile.am so that the $(XMLLINT) variable
   is used instead of hardcoding xmllint

 * Apply the patch in the host build of shared-mime-info

 * In the target build, pass an XMLLINT environment variable so that
   the xmllint binary isn't searched in $(STAGING_DIR) but in
   $(HOST_DIR)

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>