Fix Qtopia build with external toolchain
The fixes made to qmake.conf for the cross-compiling tools were not
sufficient in the case of an external toolchain, since the CFLAGS were
not taken into account. And the --sysroot option contained in the
CFLAGS are very important for external toolchains to work.

Additionally, it makes sure that Qtopia is compiled with the proper
CFLAGS, even if an external toolchain isn't used.