xorg: reorganize the selection of the X.org server type
The selection of the X.org server type was really strange. The user
had to select between none, tinyx and modular. Now, the menuconfig
interface display a single item for X.org, that can be
selected/unselected. This entry contains a submenu, that allows to
select the type of the X.org server (tinyX or modular) and that allows
(as before) to enable/disable X.org libraries, drivers, etc.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>