dillo: use select for library dependencies
Use select independs of depends on for library dependencies. We keep
the depends on for the X.org dependency so that dillo can only be
selected once X.org is enabled (which, contrary to the library
dependencies, is a pretty obvious dependency for the user). We then
get rid of the configuration interface comment.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>