webkit: select cairo PNG feature
Webkit currently depends on libgtk2, which itself depends on
Cairo. However, the dependency of libgtk2 is only on
BR2_PACKAGE_CAIRO, BR2_PACKAGE_CAIRO_PS and
BR2_PACKAGE_CAIRO_PDF. While this might be enough for libgtk2 to build
and work, Webkit makes direct use of the PNG functionnalities of
Cairo. If this functionnality is not available in Cairo, the Webkit
build complains that cairo_surface_write_to_png_stream() is not
defined.

Therefore, we make sure that WebKit selects BR2_PACKAGE_CAIRO_PNG.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>