SDL_image: fix runtime libjpeg failure
By default SDL_image loads libjpeg at runtime on-demand. The find_lib
routine in the configure script of SDL_image searches the /usr/lib directory
of the host machine to find the file name of the jpeg library.

When the host on which Buildroot is being built contains newer version of
libjpeg (e.g.  libjpeg7), the configure script incorrectly uses its file
name. The result is a runtime failure to load libjpeg.

Fix this by disabling runtime load of libjpeg.

[Peter: disable for png+tiff as well]

Signed-off-by: Baruch Siach <baruch@tkos.co.il>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>