package/mtd: Remove obsolete utilities, add new ones
A number of options under "MTD tools selection" are no longer
available.  Remove them to avoid possible build errors.

In addition, the nandtest utility is present in MTD-utils release
so remove the depends on BR2_PACKAGE_MTD_UTILS_GIT.

Add a number of new utilities that exist in the latest version.
All of these do not have a 'default y' so existing configurations
should not change.

Reorder the MTD_TARGETS_* to match the Config.in so it's easier
to notice missing utilities.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>