alsa-utils: force the variant of ncurses library
Two variants of the ncurses library exist: the normal one, and the
ncursesw one, which has support for wide char. Currently, Buildroot
only builds the normal variant (the second variant requires
--enable-widec while compiling ncurses).

Unfortunately, when libncursesw is installed on the host, the
configure script of alsa-utils finds /usr/bin/ncursesw5-config and
thinks that the target has this version. Unfortunately, as this is not
the case, it causes a configure failure when trying to link a sample
program against libpanelw (which is part of ncurses).

Therefore, we force the libncurses variant used by alsa-utils to the
normal variant.

Later on, if needed, support for the wide-char variant of the
libncurses library could be added.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>