lua: make sure that CFLAGS are passed
We were passing a custom MYCFLAGS value to Lua's Makefile, but because
of $(TARGET_CONFIGURE_OPTS), we were also passing a CFLAGS variable,
which was overriding Lua's internal CFLAGS variable. The result was
that MYCFLAGS wasn't taken into account.

Extracted from the patch proposed by François Perrad in bug #2353.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>