autotargets: don't use install-strip
Many packages using the AUTOTARGETS infrastructure do not support the
install-strip target. Moreover, we are already stripping all binary
files in the $(TARGET_DIR) anyway, so using install-strip is a bit
useless.

So, we use "make install" in all cases, and let Buildroot do the
stripping of the binaries if BR2_ENABLE_DEBUG is not set.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>