sdl_mixer: disable plugins not available in BR, add optional mp3/ogg support
Closes #2521

Disable plugins with dependencies we don't have in BR / aren't enabled, so
the configure script doesn't get confused if they are installed on the
host. Also enable ogg/libmad support if enabled in BR.

Notice: flac support is disabled for now as the sdl_mixer configure script
doesn't work when cross compiling.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>