coreutils: fixed missing hostname
hostname is no longer installed by default (ChangeLog-2007 - 2007-08-28), and
has to be enabled with --enable-install-program=hostname. hostname's man file is
not included, and because of this, during make, help2man is run against
src/hostname to generate it's man file. src/hostname will not run on the host
system, causing the build to fail.

Generated the hostname man file, after building coreutils on the host
system, and generated this patch, which fixes the build error.

Signed-off-by: Martin Banky <Martin.Banky@gmail.com>
Acked-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>