mysql_client: bump to 5.1.53 to fix abi check problem
MySQL 5.1.47 is affected by a bug in one test called the "ABI check",
which does not work with recent gcc versions. This bug is referenced
as http://bugs.mysql.com/bug.php?id=52514 in MySQL bug tracker.

Since it has been fixed in newer versions of MySQL, we simply bump the
version to the latest available in the 5.1.x series.

The patches have not changed, they have only been refreshed.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>