imagemagick: add patch to fix libxml2 issue and remove useless patch
The patch imagemagick-6.6.4-add-errno-h-if-argz-h-does-not-exist.patch
was not applied anymore due to a difference in the version number, and
it didn't prevent imagemagick to be built. It was introduced several
years ago together with the ImageMagick package itself, so presumably
it is no longer needed.

The new patch allows ImageMagick to use the correct xml2-config script
to get the proper location for XML2 headers and libraries. Otherwise,
-I/usr/include/libxml2 is found in the compile flags.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>