mplayer: fix build on i386
Mplayer needs to be built with -fomit-frame-pointer on x86, otherwise
gcc errors out on the inline asm with too few registers.

Also disable yasm support, both because this mplayer revision forgets
to build the gpl'ed asm snippets (fixed in r32736), breaking the build
- And also to ensure the build is reproducable and doesn't use host
components behind the users' back (E.G. if yasm is available on host).

Once we have added yasm as a package, this can get removed again (and
mplayer svn rev bumped) for the performance boost it gives.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>