dbus: make it easier to select an XML library (select rather than depends on)
D-Bus requires an XML library, expat or libxml2.  There already is a choice
between them in the config menu, but it requires the user to first select
one of the libraries, and then select dbus.  With this patch, the choice
automatically selects the XML library.  Note that this removes the
automatic choice of libxml2 for dbus if libxml2 had already been selected,
i.e. in that case both libxml2 and expat will be selected unless the user
takes action.

An alternative would be to remove the choice completely, and to take the
path of bluez-utils: select BR2_PACKAGE_EXPAT if !BR2_PACKAGE_DBUS_LIBXML2

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>