heirloom-mailx: run makeconfig in configure step
The heirloom-mailx Makefile runs makeconfig if this has not been done
before.  In a parallel build, this means makeconfig may be run several
times in parallel because there are two targets that depend on makeconfig.
See for instance
http://autobuild.buildroot.net/results/d24854be69961f71db189f9d804d4bd2cfa078da

To avoid this, run makeconfig in the configure step, which is not parallel.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>