socat: fix build failure due to ext2_fs.h header, bump version
In kernel >= 3.3, the ext2_fs.h can no longer be imported from
userspace. This has been fixed for internal toolchains by adding a
patch to kernel headers, but this doesn't work with toolchains
generated by Crosstool-NG, or potentially upcoming external
toolchains.

socat in fact has a test in its configure.in, but the configure was
generated too long ago, and the generated test relies on the
preprocessor result and not the compiler result (but warns that in the
future, the compiler result will be used instead of the preprocessor
result).

So, by running autoconf on this package, we fix the problem: it
properly checks whether ext2_fs is usable or not, and acts
accordingly. Of course, it means that with recent versions of the
kernel, ext2-specific features of socat are unavailable, and we'll
have to wait for the socat developers to adapt their code so that they
use the e2fsprogs headers.

We also bump the version, since a new minor version fixing a security
problem has been released.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>